Parser.Parsec

Description

Documentation

newtype Parser i e o

The parser type.

Constructors

Parser

Fields

runParser :: i -> Result i e (i, o)

satisfy :: (Input i, Token i ~ a) => (a -> Bool) -> Parser i e a

The base predicate parser. Succeeds if the predicate is true for the next token. >>> runParser (satisfy (== a)) "abc" Result ("bc",a)

char :: (Input i, Token i ~ Char) => Char -> Parser i e Char

Parses a character. >>> runParser (char a) "abc" Result ("bc",a)

digit :: (Input i, Token i ~ Char) => Parser i e Char

Parses a digit. >>> runParser digit "123" Result ("23",'1')

digits :: (Input i, Token i ~ Char) => Parser i e i

Parses a sequence of digits. >>> runParser digits "123" Result ("","123")

newline :: (Input i, Token i ~ Char) => Parser i e Char

Parses a newline character. >>> runParser newline "na" Result ("a",'n')

string :: (Input i, Token i ~ Char) => i -> Parser i e i

Parses a string. >>> runParser (string "abc") "abcdef" Result ("def","abc")

space :: (Input i, Token i ~ Char) => Parser i e Char

Parses a space character. >>> runParser space " a" Result ("a",' ')

spaces :: (Input i, Token i ~ Char) => Parser i e i

Parses a sequence of space characters. >>> runParser spaces " a" Result ("a"," ")

symbol :: (Input i, Token i ~ Char) => i -> Parser i e i

Parses a symbol. >>> runParser (symbol "abc") "abcdef" Result ("def","abc")

anySymbol :: (Input i, Token i ~ Char) => [i] -> Parser i e i

Parses any symbol from a list of symbols. >>> runParser (anySymbol ["abc", "def"]) "defghi" Result ("ghi","def")

between :: Parser i e o -> Parser i e o' -> Parser i e o'' -> Parser i e o''

Parses a value between two parsers. >>> runParser (between (char '[') (char ']') (char a)) "[a]" Result ("", a)

betweenSymbols :: (Input i, Token i ~ Char) => i -> i -> Parser i e a -> Parser i e a

Parses a value between two symbols. >>> runParser (betweenSymbols "[" "]" (char a)) "[a]" Result ("", a)

brackets :: (Input i, Token i ~ Char) => Parser i e a -> Parser i e a

Parses a value between brackets. >>> runParser (brackets (char a)) "[a]" Result ("", a)

braces :: (Input i, Token i ~ Char) => Parser i e a -> Parser i e a

Parses a value between braces. >>> runParser (braces (char a)) "{a}" Result ("", a)

parens :: (Input i, Token i ~ Char) => Parser i e a -> Parser i e a

Parses a value between parentheses. >>> runParser (parens (char a)) "(a)" Result ("", a)

ignoreSpaces :: (Input i, Token i ~ Char) => Parser i e a -> Parser i e a

Parses a value between spaces. >>> runParser (ignoreSpaces (char a)) " a" Result ("", a)

letter :: (Input i, Token i ~ Char) => Parser i e Char

Parses a letter. >>> runParser letter "abc" Result ("bc",a)

letters :: (Input i, Token i ~ Char) => Parser i e i

Parses a sequence of letters. >>> runParser letters "abc" Result ("","abc")

eof :: Input i => Parser i e ()

Parses the end of the input. >>> runParser eof "" Result ("",())

lookahead :: (Input i, Token i ~ Char) => Parser i i (Maybe Char)

Parses the next token without consuming it. >>> runParser lookahead "a" Result ("a",Just a)

data Result i e a

The result type for the parser.

Constructors

Errors [Error i e]

Result a

data Error :: Type -> Type -> Type where

The error type for the parser.

Constructors

EndOfInput :: Error i e
The input ended unexpectedly.

Unexpected :: i -> Error i e
An unexpected token was encountered.

class (Eq i, IsString i, Monoid i) => Input (i :: Type) where

The Input type class for parsing

Associated Types

type Token i

The token type of the input.

Methods

uncons :: i -> Maybe (Token i, i)

Uncons the input.

head :: i -> Maybe (Token i)

Get the head of the input.

null :: i -> Bool

Check if the input is empty.

unpack :: i -> String

Unpack the input to a string.

chainl1 :: Parser i e a -> Parser i e (a -> a -> a) -> Parser i e a

Parses a left-associative chain of values. >>> runParser (chainl1 (char a) (char +)) "a+a+a" Result ("","+")

manyTill :: Parser i e a -> Parser i e b -> Parser i e [a]

Parses a list of values until the end parser succeeds. >>> runParser (manyTill (char a) (char b)) "aaab" Result ("b","aaa")

someTill :: Parser i e a -> Parser i e b -> Parser i e [a]

Parses a list of values until the end parser succeeds. >>> runParser (someTill (char a) (char b)) "aaab" Result ("b","aaa")

try :: Parser i e a -> Parser i e a

Tries to parse a value. If the parser fails, it returns an empty result. >>> runParser (try (char a)) "b" Errors []

spaceNoNewline :: (Input i, Token i ~ Char) => Parser i e Char

Parses a space character that is not a newline. >>> runParser spaceNoNewline " a" Result ("a",' ')

spacesNoNewline :: (Input i, Token i ~ Char) => Parser i e i

Parses a sequence of space characters that are not newlines. >>> runParser spacesNoNewline " a" Result ("a"," ")

many :: Alternative f => f a -> f [a]

Zero or more.

some :: Alternative f => f a -> f [a]

One or more.

optional :: Alternative f => f a -> f (Maybe a)

One or none.

It is useful for modelling any computation that is allowed to fail.

Examples

Expand

Using the Alternative instance of Control.Monad.Except, the following functions:

>>> import Control.Monad.Except

>>> canFail = throwError "it failed" :: Except String Int
>>> final = return 42                :: Except String Int

Can be combined by allowing the first function to fail:

>>> runExcept $ canFail *> final
Left "it failed"
>>> runExcept $ optional canFail *> final
Right 42