Problem

Description

Documentation

class Problem (a :: Type) where

Represents a problem that can be solved by the SAT solver.

Associated Types

type Variable a

The type of variables in the problem.

Methods

toDIMACS :: a -> DIMACS

Converts the problem to DIMACS format.

decode :: a -> Solutions -> a

Decodes the solution to the problem.

encodeVar :: a -> Variable a -> Int

Encodes a variable to an integer.

example :: a

Parses a problem from a string.

parse :: Text -> Maybe a

Parses a problem from a string.

parseFile :: Problem a => Text -> IO (Maybe a)

Parses a problem from a file.

solve :: Problem a => a -> Maybe a

Solves a problem. Returns Nothing if the problem is unsatisfiable. Returns Just the solution if the problem is satisfiable.

>>> solve Sudoku.sudokuSixteen
Just ...

>>> solve Nonogram.exampleNonogram
Just ...

>>> solve DIMACS.exampleDIMACS
Just ...

>>> solve (SAT.toCNF (SAT.fromExpr (SAT.Var 1)))
Just 1

>>> solve (SAT.toCNF (SAT.fromExpr (SAT.Not (SAT.Var 1)) `SAT.And` SAT.fromExpr (SAT.Var 1)))
Nothing

toExpr :: Problem a => a -> Expr Int

Converts a problem to an expression.

>>> toExpr Sudoku.sudokuSixteen
...

>>> toExpr Nonogram.exampleNonogram
...

>>> toExpr DIMACS.exampleDIMACS
...

>>> toExpr (SAT.toCNF (SAT.fromExpr (SAT.Var 1)))
SAT.Var 1

toCNF :: Problem a => a -> CNF

Converts a problem to CNF.

>>> toCNF Sudoku.sudokuSixteen
...

>>> toCNF Nonogram.exampleNonogram
...

>>> toCNF DIMACS.exampleDIMACS
...

>>> toCNF (SAT.toCNF (SAT.fromExpr (SAT.Var 1)))
CNF [[1]]

>>> toCNF (SAT.toCNF (SAT.fromExpr (SAT.Not (SAT.Var 1)) `SAT.And` SAT.fromExpr (SAT.Var 1)))
CNF [[-1][1]]

isSatisfiable :: Problem a => a -> Bool

Checks if a problem is satisfiable.

>>> isSatisfiable Sudoku.sudokuSixteen
True

>>> isSatisfiable Nonogram.exampleNonogram
True

>>> isSatisfiable DIMACS.exampleDIMACS
True

>>> isSatisfiable (SAT.toCNF (SAT.fromExpr (SAT.Var 1)))
True

>>> isSatisfiable (SAT.toCNF (SAT.fromExpr (SAT.Not (SAT.Var 1)) `SAT.And` SAT.fromExpr (SAT.Var 1)))
False

getNumClauses :: Problem a => a -> Integer

Gets the number of clauses in a problem.

>>> getNumClauses Sudoku.sudokuSixteen
...

getNumVars :: Problem a => a -> Integer

solveWith :: Problem a => (Expr Int -> Maybe Solutions) -> a -> Maybe a

averageClauseLength :: Problem a => a -> Double

getTotalNumLiterals :: Problem a => a -> Integer