SAT.CNF

Description

Documentation

applyLaws :: Eq a => Expr a -> Expr a

Applies the laws of distribution and De Morgan to an expression.

>>> applyLaws (Or (And (Var 1) (Var 2)) (Not (Var 3)))
And (Or (Not (Var 3)) (Var 1)) (Or (Not (Var 3)) (Var 2))

toCNF :: Expr Int -> CNF

Converts an expression to CNF.

>>> toCNF (Or (And (Var 1) (Var 2)) (Not (Var 3)))
CNF [[-3,1],[-3,2]]

newtype CNF

The CNF type.

Constructors

CNF [Clause]

type Clause = [Literal]

The clause type.

type Literal = Int

The literal type.

addClause :: CNF -> Clause -> CNF

Adds a clause to a CNF.

>>> addClause (CNF [[1,2]]) [3]
CNF [[3],[1,2]]

type Assignment = IntMap Bool

type DecisionLevel = Int

isNegative :: Literal -> Bool

Checks if a literal is negative.

>>> isNegative (-1)
True

>>> isNegative 1
False

isNegative x == (x < 0)

tseitin :: Expr Int -> State Int (CNF, Literal)

toTseitin :: Expr Int -> CNF

varOfLiteral :: Literal -> Int

Gets the variable of a literal.

>>> varOfLiteral 1
1

>>> varOfLiteral (-1)
1

varValue :: Assignment -> Int -> Maybe Bool

Returns the value of a variable in an assignment.

literalValue :: Assignment -> Literal -> Maybe Bool

Returns the value of a literal in an assignment.

negateLiteral :: Literal -> Literal

initAssignment :: IntSet -> Assignment