SAT.Optimisers

Description

Documentation

unitPropagate :: CNF -> Assignment -> DecisionLevel -> (Assignment, CNF)

Propagates a unit clause in a CNF.

>>> unitPropagate (CNF [[1, 2], [-2, -3], [3, 4]]) IntMap.empty 0
(CNF {clauses = [[1,2],[-2,-3],[3,4]]},fromList [])

>>> unitPropagate (CNF [[1], [2, 3], [3, 4]]) IntMap.empty 0
(CNF {clauses = [[2,3],[3,4]]},fromList [(1,True)])

substitute :: Literal -> Bool -> CNF -> CNF

Substitutes a literal in a CNF.

>>> substitute 1 True (CNF [[1, 2], [2, 3], [3, 4]])
CNF {clauses = [[2,3],[3,4]]}

collectLiterals :: CNF -> [Int]

Collects all literals in a CNF.

>>> collectLiterals (CNF [[1, 2], [2, 3], [3, 4]])
[1,2,2,3,3,4]

collectLiteralsToSet :: CNF -> IntSet

Collects all literals in a CNF and returns them as a set.

>>> collectLiteralsToSet (CNF [[1, 2], [2, 3], [3, 4]])
fromList [1,2,3,4]

uniqueOnly :: forall a. Ord a => [a] -> [a]

Removes duplicates from a list.

>>> uniqueOnly [1, 2, 3, 2, 1]
[1,2,3]

eliminateLiterals :: CNF -> Assignment -> (Assignment, CNF)

eliminates any literals that are either positive or negative from the CNF and updates the assignment returns the updated CNF and assignment

assign :: Assignment -> Literal -> Bool -> Assignment

Assigns a value to a literal.

>>> assign IntMap.empty 1 True 0
fromList [(1,True)]

partialAssignment :: Assignment -> CNF -> CNF

Applies a partial assignment to a CNF.

>>> partialAssignment (IntMap.fromList [(1, (True, 0))]) (CNF [[1, 2], [-2, -3], [3, 4]])
CNF {clauses = [[2],[-3],[3,4]]}

pickVariable :: VSIDS -> Maybe (Literal, VSIDS)

Picks a variable.

>>> pickVariable (IntMap.fromList [(1, 1), (2, 2), (3, 3)])
Just (3,fromList [(1,1.0),(2,2.0)])

pickVariableM :: SolverM (Maybe Literal)

literalPolarities :: CNF -> IntMap Polarity

Finds the polarities of the literals in a CNF.

>>> literalPolarities (CNF [[1, 2], [-2, -3], [3, 4]])
fromList [(1,Positive),(2,Mixed),(3,Mixed),(4,Positive)]

decayM :: SolverM ()

eliminateLiteralsM :: SolverM ()

findUnitClause :: CNF -> Maybe (Int, Bool)

Finds a unit clause in a CNF (a clause with only one literal).

>>> findUnitClause (CNF [[1, 2], [-2, -3], [3, 4]])
Nothing

>>> findUnitClause (CNF [[1], [2, 3], [3, 4]])
Just (1,True)

unitPropagateM :: SolverM (Maybe Clause)

assignM :: Literal -> Bool -> SolverM ()

removeTautologies :: CNF -> CNF

Removes tautologies from a CNF. A tautology is a clause that contains both a literal and its negation, and is always true therefore. This function removes all such clauses from the CNF.

>>> removeTautologies (CNF [[-2, 2], [-2, -3], [3, 4]])
CNF {[[-2,-3],[3,4]]}

isSat :: CNF -> Bool

Checks if a CNF is satisfied

>>> isSat (CNF [[1, 2], [2, 3], [3, 4]])
False

>>> isSat (CNF [])
True

isUnsat :: CNF -> Bool

Checks if a CNF is unsatisfiable

>>> isUnsat (CNF [[1, 2], [2, 3], [3, 4]])
False

>>> isUnsat (CNF [[]])
True

backtrack :: DecisionLevel -> SolverM ()

Backtracks to a given decision level.

analyseConflict :: Clause -> SolverM (Clause, DecisionLevel)

addDecision :: Literal -> SolverM ()

addPropagation :: Literal -> Clause -> SolverM ()

isSatisfied :: SolverM Bool

adjustScoresM :: Clause -> SolverM ()

getClauseStatus :: Clause -> SolverM ClauseStatus

findCanditateWatchedLiteral :: Clause -> SolverM (Maybe Literal)

data ClauseStatus

Constructors

Unit Literal

UNSAT

SAT

Unresolved

pickLiteralM :: SolverM Literal