SAT.Solver

Description

simplify for now, don't use partials, don't worry about inefficiencies that is what watch literals aim to solve or i could have my partial list of maybe clause, which i can index on the main clauses to get them but must compare that to watched literals when i implement that watched literals can also be checked for tautolofy should wathced literals be stored as the abs value or either ?

so for now, when finding units just compare the main cnf against the assignment follow the tutorial (but with ig) experiment with the use of pure literals, not sure how they fit in with the watched literals (maybe tautlogies could be useful too)

maybe worth dragging the cnf around in state to add learned clauses iwht little hassle consider not using vsids for now, just bare minimum biggest issue is that i need the conflicting clause to learn from it, hence lack of partials another issue is if my solver is dramatically more inefficient due to partials lacking, how can i know if cdcl is working ?? i suppose i check the learned clauses against other solvers

simplify simplify simpify

issue with using partial assignment to check for isSat is learned clauses

Documentation

satisfiable :: CNF -> Bool

Checks if a CNF is satisfiable.

type Solutions = IntSet

The solutions type.

checkValue :: Solutions -> Literal -> Bool

Checks if a value is in the solutions.

>>> checkValue (IntSet.fromList [1, 2, 3]) 2
True

findFreeVariable :: CNF -> Maybe Literal

Finds a free variable at random.

getSolutions :: CNF -> Maybe Solutions

bruteForce :: Expr Int -> Maybe Solutions

withUnitPropagation :: Expr Int -> Maybe Solutions

withPureLiteralElimination :: Expr Int -> Maybe Solutions

withPureLiteralAndUnitPropagation :: Expr Int -> Maybe Solutions

pureLitOnlyAsPreprocess :: Expr Int -> Maybe Solutions

literalEvery100 :: Expr Int -> Maybe Solutions

withTautologyElimination :: Expr Int -> Maybe Solutions